move booklist component
