+++ src/app/components/BookList/components/BookCard/index.tsx
