+++ src/app/components/BookList/components/BookCard/styles.ts
